refactor: rename data-value-mutations dir
