+++ src/data-workspace/inputs/generic-input.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React, { useState } from 'react'
 import { useField } from 'react-final-form'
 import { VALUE_TYPES } from '../../shared/index.js'
-import { useSetDataValueMutation } from '../use-data-value-mutation/index.js'
+import { useSetDataValueMutation } from '../data-value-mutations/index.js'
 import styles from './inputs.module.css'
 import { InputPropTypes } from './utils.js'
 import { validatorsByValueType } from './validators.js'
